@@ -3,12 +3,12 @@ description: Demonstrates how to use the just_audio plugin.
 publish_to: 'none'
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.14.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  audio_session: ^0.1.7
+  audio_session: ^0.1.14
   rxdart: ^0.27.2
   just_audio:
     path: ../../just_audio
