@@ -1,14 +1,19 @@
 name: just_audio_background
 description: An add-on for just_audio that supports background playback and media notifications.
 homepage: https://github.com/ryanheise/just_audio/tree/master/just_audio_background
-version: 0.0.1-beta.8
+version: 0.0.1-beta.10
+topics:
+  - audio
+  - sound
+  - player
+  - background
 
 dependencies:
-  just_audio_platform_interface: ^4.2.0
+  just_audio_platform_interface: ^4.2.1
   # just_audio_platform_interface:
   #   path: ../just_audio_platform_interface
-  audio_service: ^0.18.6
-  audio_session: ^0.1.7
+  audio_service: ^0.18.9
+  audio_session: ^0.1.14
   flutter:
     sdk: flutter
   flutter_web_plugins:
@@ -20,5 +25,5 @@ dev_dependencies:
   flutter_lints: ^2.0.1
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=1.12.13+hotfix.5"
+  sdk: ">=2.14.0 <4.0.0"
+  flutter: ">=3.0.0"
