@@ -86,13 +86,13 @@ class AudioPlayer {
   /// implementation. When switching between active and inactive modes, this is
   /// used to cancel the subscription to the previous platform's events and
   /// subscribe to the new platform's events.
-  StreamSubscription? _playbackEventSubscription;
+  StreamSubscription<PlaybackEventMessage>? _playbackEventSubscription;
 
   /// The subscription to the data event channel of the current platform
   /// implementation. When switching between active and inactive modes, this is
   /// used to cancel the subscription to the previous platform's events and
   /// subscribe to the new platform's events.
-  StreamSubscription? _playerDataSubscription;
+  StreamSubscription<PlayerDataMessage>? _playerDataSubscription;
 
   final String _id;
   final _proxy = _ProxyHttpServer();
@@ -142,13 +142,14 @@ class AudioPlayer {
 
   /// Creates an [AudioPlayer].
   ///
-  /// If [userAgent] is specified, it will be included in the header of all HTTP
-  /// requests on Android, iOS and macOS to identify your agent to the server.
-  /// If set, just_audio will create a cleartext local HTTP proxy on your device
-  /// to forward HTTP requests with headers included. If [userAgent] is not
-  /// specified, this will default to Apple's Core Audio user agent on iOS/macOS
-  /// and to just_audio's own user agent on Android. On Web, the browser will
-  /// override any specified user-agent string with its own.
+  /// Apps requesting remote URLs should specify a `[userAgent]` string with
+  /// this constructor which will be included in the `user-agent` header on all
+  /// HTTP requests (except on web where the browser's user agent will be sent).
+  /// This header helps to identify to the server which app is submitting the
+  /// request. If unspecified, it will default to Apple's Core Audio user agent
+  /// on iOS/macOS, or just_audio's user agent on Android. Note: this feature
+  /// is implemented via a local HTTP proxy which requires non-HTTPS support to
+  /// be enabled. See the README page for setup instructions.
   ///
   /// The player will automatically pause/duck and resume/unduck when audio
   /// interruptions occur (e.g. a phone call) or when headphones are unplugged.
@@ -614,8 +615,8 @@ class AudioPlayer {
     }
 
     Timer? currentTimer;
-    StreamSubscription? durationSubscription;
-    StreamSubscription? playbackEventSubscription;
+    StreamSubscription<Duration?>? durationSubscription;
+    StreamSubscription<PlaybackEvent>? playbackEventSubscription;
     void yieldPosition(Timer timer) {
       if (controller.isClosed) {
         timer.cancel();
@@ -1992,13 +1993,13 @@ class _ProxyHttpServer {
   String _requestKey(Uri uri) => '${uri.path}?${uri.query}';
 
   /// Start the server if it is not already running.
-  Future ensureRunning() async {
+  Future<dynamic> ensureRunning() async {
     if (_running) return;
     return await start();
   }
 
   /// Starts the server.
-  Future start() async {
+  Future<dynamic> start() async {
     _running = true;
     _server = await HttpServer.bind(InternetAddress.loopbackIPv4, 0);
     _server.listen((request) async {
@@ -2015,7 +2016,7 @@ class _ProxyHttpServer {
   }
 
   /// Stops the server
-  Future stop() async {
+  Future<dynamic> stop() async {
     if (!_running) return;
     _running = false;
     return await _server.close();
@@ -2115,7 +2116,7 @@ abstract class AudioSource {
   /// AudioSource.uri(Uri.file(filePath));
   /// ```
   static UriAudioSource file(String filePath, {dynamic tag}) {
-    return AudioSource.uri(Uri.file(filePath));
+    return AudioSource.uri(Uri.file(filePath), tag: tag);
   }
 
   /// Convenience method to create an audio source for an asset.
@@ -2132,7 +2133,7 @@ abstract class AudioSource {
       {String? package, dynamic tag}) {
     final keyName =
         package == null ? assetPath : 'packages/$package/$assetPath';
-    return AudioSource.uri(Uri.parse('asset:///$keyName'));
+    return AudioSource.uri(Uri.parse('asset:///$keyName'), tag: tag);
   }
 
   AudioSource() : _id = _uuid.v4();
@@ -2841,7 +2842,7 @@ class LockCachingAudioSource extends StreamAudioSource {
     final mimeFile = await _mimeFile;
     await mimeFile.writeAsString(mimeType);
     final inProgressResponses = <_InProgressCacheResponse>[];
-    late StreamSubscription subscription;
+    late StreamSubscription<List<int>> subscription;
     var percentProgress = 0;
     void updateProgress(int newPercentProgress) {
       if (newPercentProgress != percentProgress) {
@@ -3197,6 +3198,8 @@ _ProxyHandler _proxyHandlerForUri(
         // TODO: Handle other playlist formats similarly?
         final m3u8 = await originResponse.transform(utf8.decoder).join();
         for (var line in const LineSplitter().convert(m3u8)) {
+          line = line.replaceAllMapped(
+              RegExp(r'#EXT-X-MEDIA:.*?URI="(.*?)".*'), (m) => m[1]!);
           line = line.replaceAll(RegExp(r'#.*$'), '').trim();
           if (line.isEmpty) continue;
           try {
