@@ -132,7 +132,7 @@ class _JustAudioBackgroundPlugin extends JustAudioPlatform {
               "just_audio_background supports only a single player instance");
     }
     _player = _JustAudioPlayer(
-      id: request.id,
+      initRequest: request,
     );
     return _player!;
   }
@@ -157,16 +157,16 @@ class _JustAudioBackgroundPlugin extends JustAudioPlatform {
 }
 
 class _JustAudioPlayer extends AudioPlayerPlatform {
+  final InitRequest initRequest;
   final eventController = StreamController<PlaybackEventMessage>.broadcast();
   final playerDataController = StreamController<PlayerDataMessage>.broadcast();
   bool? _playing;
-  int? _index;
   IcyMetadataMessage? _icyMetadata;
   int? _androidAudioSessionId;
   late final _PlayerAudioHandler _playerAudioHandler;
 
-  _JustAudioPlayer({required String id}) : super(id) {
-    _playerAudioHandler = _PlayerAudioHandler(id);
+  _JustAudioPlayer({required this.initRequest}) : super(initRequest.id) {
+    _playerAudioHandler = _PlayerAudioHandler(initRequest);
     _audioHandler.inner = _playerAudioHandler;
     _audioHandler.playbackState.listen((playbackState) {
       broadcastPlaybackEvent();
@@ -180,11 +180,6 @@ class _JustAudioPlayer extends AudioPlayerPlatform {
         case 'androidAudioSessionId':
           _androidAudioSessionId = event['value'] as int?;
           broadcastPlaybackEvent();
-          break;
-        case 'currentIndex':
-          _index = event['value'] as int?;
-          // The event is broadcast in response to the next mediaItem update
-          // which happens immediately after this.
           break;
       }
     });
@@ -218,7 +213,7 @@ class _JustAudioPlayer extends AudioPlayerPlatform {
       bufferedPosition: playbackState.bufferedPosition,
       icyMetadata: _icyMetadata,
       duration: _playerAudioHandler.currentMediaItem?.duration,
-      currentIndex: _index,
+      currentIndex: playbackState.queueIndex,
       androidAudioSessionId: _androidAudioSessionId,
     ));
     if (playbackState.playing != _playing) {
@@ -238,9 +233,8 @@ class _JustAudioPlayer extends AudioPlayerPlatform {
       playerDataController.stream;
 
   @override
-  Future<LoadResponse> load(LoadRequest request) async {
-    return _playerAudioHandler.customLoad(request);
-  }
+  Future<LoadResponse> load(LoadRequest request) =>
+      _playerAudioHandler.customLoad(request);
 
   @override
   Future<PlayResponse> play(PlayRequest request) async {
@@ -285,41 +279,35 @@ class _JustAudioPlayer extends AudioPlayerPlatform {
       _playerAudioHandler.customSetShuffleOrder(request);
 
   @override
-  Future<SeekResponse> seek(SeekRequest request) async {
-    return await _playerAudioHandler.customPlayerSeek(request);
-  }
+  Future<SeekResponse> seek(SeekRequest request) =>
+      _playerAudioHandler.customPlayerSeek(request);
 
   @override
   Future<ConcatenatingInsertAllResponse> concatenatingInsertAll(
-      ConcatenatingInsertAllRequest request) async {
-    return _playerAudioHandler.customConcatenatingInsertAll(request);
-  }
+          ConcatenatingInsertAllRequest request) =>
+      _playerAudioHandler.customConcatenatingInsertAll(request);
 
   @override
   Future<ConcatenatingRemoveRangeResponse> concatenatingRemoveRange(
-      ConcatenatingRemoveRangeRequest request) async {
-    return await _playerAudioHandler.customConcatenatingRemoveRange(request);
-  }
+          ConcatenatingRemoveRangeRequest request) =>
+      _playerAudioHandler.customConcatenatingRemoveRange(request);
 
   @override
   Future<ConcatenatingMoveResponse> concatenatingMove(
-      ConcatenatingMoveRequest request) async {
-    return await _playerAudioHandler.customConcatenatingMove(request);
-  }
+          ConcatenatingMoveRequest request) =>
+      _playerAudioHandler.customConcatenatingMove(request);
 
   @override
   Future<SetAndroidAudioAttributesResponse> setAndroidAudioAttributes(
-      SetAndroidAudioAttributesRequest request) async {
-    return await _playerAudioHandler.customSetAndroidAudioAttributes(request);
-  }
+          SetAndroidAudioAttributesRequest request) =>
+      _playerAudioHandler.customSetAndroidAudioAttributes(request);
 
   @override
   Future<SetAutomaticallyWaitsToMinimizeStallingResponse>
       setAutomaticallyWaitsToMinimizeStalling(
-          SetAutomaticallyWaitsToMinimizeStallingRequest request) async {
-    return await _playerAudioHandler
-        .customSetAutomaticallyWaitsToMinimizeStalling(request);
-  }
+              SetAutomaticallyWaitsToMinimizeStallingRequest request) =>
+          _playerAudioHandler
+              .customSetAutomaticallyWaitsToMinimizeStalling(request);
 }
 
 class _PlayerAudioHandler extends BaseAudioHandler
@@ -357,12 +345,12 @@ class _PlayerAudioHandler extends BaseAudioHandler
 
   List<MediaItem>? get currentQueue => queue.nvalue;
 
-  _PlayerAudioHandler(String playerId) {
-    _init(playerId);
+  _PlayerAudioHandler(InitRequest initRequest) {
+    _init(initRequest);
   }
 
-  Future<void> _init(String playerId) async {
-    final player = await _platform.init(InitRequest(id: playerId));
+  Future<void> _init(InitRequest initRequest) async {
+    final player = await _platform.init(initRequest);
     _playerCompleter.complete(player);
     final playbackEventMessageStream = player.playbackEventMessageStream;
     playbackEventMessageStream.listen((event) {
@@ -411,10 +399,6 @@ class _PlayerAudioHandler extends BaseAudioHandler
                   currentQueue![index!].copyWith(duration: track.duration);
               queue.add(currentQueue!);
             }
-            customEvent.add({
-              'type': 'currentIndex',
-              'value': track.index,
-            });
             mediaItem.add(currentMediaItem!);
           }
         });
@@ -443,13 +427,11 @@ class _PlayerAudioHandler extends BaseAudioHandler
     return LoadResponse(duration: response.duration);
   }
 
-  Future<SetVolumeResponse> customSetVolume(SetVolumeRequest request) async {
-    return await (await _player).setVolume(request);
-  }
+  Future<SetVolumeResponse> customSetVolume(SetVolumeRequest request) async =>
+      await (await _player).setVolume(request);
 
-  Future<SeekResponse> customPlayerSeek(SeekRequest request) async {
-    return await (await _player).seek(request);
-  }
+  Future<SeekResponse> customPlayerSeek(SeekRequest request) async =>
+      await (await _player).seek(request);
 
   Future<SetShuffleOrderResponse> customSetShuffleOrder(
       SetShuffleOrderRequest request) async {
@@ -493,18 +475,16 @@ class _PlayerAudioHandler extends BaseAudioHandler
   }
 
   Future<SetAndroidAudioAttributesResponse> customSetAndroidAudioAttributes(
-      SetAndroidAudioAttributesRequest request) async {
-    return await (await _player).setAndroidAudioAttributes(request);
-  }
+          SetAndroidAudioAttributesRequest request) async =>
+      await (await _player).setAndroidAudioAttributes(request);
 
   Future<SetAutomaticallyWaitsToMinimizeStallingResponse>
       customSetAutomaticallyWaitsToMinimizeStalling(
-          SetAutomaticallyWaitsToMinimizeStallingRequest request) async {
-    return await (await _player)
-        .setAutomaticallyWaitsToMinimizeStalling(request);
-  }
+              SetAutomaticallyWaitsToMinimizeStallingRequest request) async =>
+          await (await _player)
+              .setAutomaticallyWaitsToMinimizeStalling(request);
 
-  Future<void> _updateQueue() async {
+  void _updateQueue() {
     queue.add(sequence.map((source) => source.tag as MediaItem).toList());
   }
 
@@ -689,14 +669,14 @@ class _PlayerAudioHandler extends BaseAudioHandler
     }
   }
 
-  Future<void> _broadcastStateIfActive() async {
+  void _broadcastStateIfActive() {
     if (_justAudioEvent.processingState != ProcessingStateMessage.idle) {
       _broadcastState();
     }
   }
 
   /// Broadcasts the current state to all clients.
-  Future<void> _broadcastState() async {
+  void _broadcastState() {
     final controls = [
       if (hasPrevious) MediaControl.skipToPrevious,
       if (_playing) MediaControl.pause else MediaControl.play,
@@ -713,7 +693,7 @@ class _PlayerAudioHandler extends BaseAudioHandler
       androidCompactActionIndices: List.generate(controls.length, (i) => i)
           .where((i) => controls[i].action != MediaAction.stop)
           .toList(),
-      processingState: {
+      processingState: const {
         ProcessingStateMessage.idle: AudioProcessingState.idle,
         ProcessingStateMessage.loading: AudioProcessingState.loading,
         ProcessingStateMessage.buffering: AudioProcessingState.buffering,
@@ -743,7 +723,7 @@ class _Seeker {
     this.duration,
   );
 
-  void start() async {
+  Future<void> start() async {
     _running = true;
     while (_running) {
       Duration newPosition = handler.currentPosition + positionInterval;
@@ -816,14 +796,14 @@ extension AudioSourceExtension on AudioSourceMessage {
     if (self is ConcatenatingAudioSourceMessage) {
       var offset = 0;
       final childIndicesList = <List<int>>[];
-      for (var child in self.children) {
+      for (final child in self.children) {
         final childIndices =
             child.shuffleIndices.map((i) => i + offset).toList();
         childIndicesList.add(childIndices);
         offset += childIndices.length;
       }
       final indices = <int>[];
-      for (var index in self.shuffleOrder) {
+      for (final index in self.shuffleOrder) {
         indices.addAll(childIndicesList[index]);
       }
       return indices;
@@ -837,6 +817,7 @@ extension AudioSourceExtension on AudioSourceMessage {
   }
 }
 
+@immutable
 class TrackInfo {
   final int? index;
   final Duration? duration;
@@ -848,7 +829,7 @@ class TrackInfo {
       other is TrackInfo && index == other.index && duration == other.duration;
 
   @override
-  int get hashCode => "$index,$duration".hashCode;
+  int get hashCode => Object.hash(index, duration);
 
   @override
   String toString() => '($index, $duration)';
